+++ packages/plugins/openai-chatgpt/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @aio-proxy/plugin-openai-chatgpt
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.2
+  - @aio-proxy/types@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.1
+  - @aio-proxy/types@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.0
+  - @aio-proxy/types@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes