+++ packages/plugins/xai-grok/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @aio-proxy/plugin-xai-grok
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes