+++ packages/server/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @aio-proxy/server
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [[`39d1b19`](https://github.com/aio-proxy/aio-proxy/commit/39d1b1927055fa483c9d09d82b6e5e76100eee95)]:
+  - @aio-proxy/i18n@0.5.2
+  - @aio-proxy/core@0.5.2
+  - @aio-proxy/logger@0.5.2
+  - @aio-proxy/plugin-sdk@0.5.2
+  - @aio-proxy/types@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [[`1a525e8`](https://github.com/aio-proxy/aio-proxy/commit/1a525e861a0ef77668c3321f75171bb9e2880e9f)]:
+  - @aio-proxy/core@0.5.1
+  - @aio-proxy/i18n@0.5.1
+  - @aio-proxy/logger@0.5.1
+  - @aio-proxy/plugin-sdk@0.5.1
+  - @aio-proxy/types@0.5.1
+
+## 0.5.0
+
+### Minor Changes
+
+- [#125](https://github.com/aio-proxy/aio-proxy/pull/125) [`7856451`](https://github.com/aio-proxy/aio-proxy/commit/7856451f2434912a619e1c72aca44a1ccd1aaf43) Thanks [@baranwang](https://github.com/baranwang)! - server: return real upstream token counts for `/v1/messages/count_tokens` when a same-protocol raw provider is configured, and replace the `bytes/64` fallback with a character-class-weighted estimator
+
+### Patch Changes
+
+- Updated dependencies [[`c6ecfc0`](https://github.com/aio-proxy/aio-proxy/commit/c6ecfc0dc81e6cb0f0c5cd7b27b79f32cfb0955c), [`d95834a`](https://github.com/aio-proxy/aio-proxy/commit/d95834ad85ea0352f5c389497ea008c687a80d64)]:
+  - @aio-proxy/core@0.5.0
+  - @aio-proxy/i18n@0.5.0
+  - @aio-proxy/logger@0.5.0
+  - @aio-proxy/plugin-sdk@0.5.0
+  - @aio-proxy/types@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes